Codechange: [Network] Do not leak os_abstraction.h via fios.h
