+++ src/fios.h
@@ -13,7 +13,7 @@
 #include "gfx_type.h"
 #include "company_base.h"
 #include "newgrf_config.h"
-#include "network/core/tcp_content.h"
+#include "network/core/tcp_content_type.h"
 
 
 /** Special values for save-load window for the data parameter of #InvalidateWindowData. */
